Create endpoints folder to keep things organized
