+++ src/index.ts
@@ -1,3 +1,3 @@
-export * from './app-details';
 export * from './common/consts';
-export * from './top-charts';
+export * from './endpoints/app-details';
+export * from './endpoints/top-charts';
